feat: merge and streamline the demo and sandbox branch
