+++ .env.example
@@ -1,24 +1,17 @@
 # This file is a "template" of which env vars need to be defined for your application.
 # Copy this file as .env in the project root folder, and replace the values with your own ones
-# This version contains the minimal set of variables required to run your own YAFFA instance.
-# For a full list of available variables, see the .env.example.complete file
+# This version contains all possible settings and their descriptions,
+# also indicating the minimum set of variables required to run your own YAFFA instance.
 
+###############################
 ### Basic app configuration ###
+###############################
 
 # Encryption key used to encrypt session and other data
 # You can use the "php artisan key:generate" command to generate a random key, which will overwrite this value
 # Make sure to keep this key secret, and don't share it with anyone.
 # Also, don't change it after you have started using the application.
 APP_KEY=PleaseReplaceThisWithARandomStringOf32Chars
-
-# Set this to true if you need to see detailed error messages
-APP_DEBUG=false
-
-# You can specify the log channel to be used for logging
-LOG_CHANNEL=stack
-
-# The minimum log level to be written to the log files
-LOG_LEVEL=error
 
 ### Database configuration ###
 DB_CONNECTION=mysql
@@ -29,7 +22,25 @@ DB_USERNAME=yaffa_user
 DB_PASSWORD=yaffa_password
 DB_SOCKET=
 
-# Define the configuration for sending emails
+### Application behavior and settings ###
+
+# The maximum number of users that can be registered in the system.
+# For a personal instance, you can set this to 1 to prevent other people from registering.
+# Increase this value if you want to allow other users to register, or set it to empty to allow unlimited registrations.
+REGISTERED_USER_LIMIT=1
+
+# Do new users have to verify their email address before they can log in?
+# For a local, personal instance, without access to outgoing emails, this should be set to false.
+# For a public instance, this should probably be set to true, but you need to configure the email settings below according to your server setup.
+EMAIL_VERIFICATION_REQUIRED=FALSE
+
+# The email address of the administrator of the system, who will receive notifications about new users, logins, etc.
+# Leave this empty if you don't want to receive any notifications
+ADMIN_EMAIL=
+
+### Define the configuration for sending emails ###
+# This is needed if you want your users to verify their email addresses,
+# or if you want to receive notifications about new users, etc.
 MAIL_MAILER=log
 MAIL_HOST=
 MAIL_PORT=
@@ -38,6 +49,14 @@ MAIL_PASSWORD=
 MAIL_ENCRYPTION=null
 MAIL_FROM_ADDRESS=info@yaffa.test
 MAIL_FROM_NAME=YAFFA
+
+# The URL of the application, used for generating links in emails
+# This URL is independent of the actual URL used to access the application, which you can configure in the web server
+APP_URL=http://localhost
+
+#############################
+### Advanced app features ###
+#############################
 
 ### Configuration of external services ###
 # Your key for the Alpha Vantage API, which is used to fetch stock prices
@@ -51,26 +70,9 @@ MAILBOX_HTTP_PASSWORD=
 INCOMING_RECEIPTS_EMAIL=
 
 # Your OpenAI key, which is used when AI is asked to process incoming email receipts
+# Note: this will be used by all users, so make sure to use a key that has enough capacity or billing cap to prevent abuse
+# Related note: this should be moved to a per-user setting in the future
 OPENAI_API_KEY=
-
-### Application-specific configuration ###
-
-# The maximum number of users that can be registered in the system.
-# For a personal instance, you can set this to 1 to prevent other people from registering.
-# Increase this value if you want to allow other users to register, or set it to empty to allow unlimited registrations.
-REGISTERED_USER_LIMIT=1
-
-# Do new users have to verify their email address before they can log in?
-# For a local, personal instance, without access to outgoing emails, this should be set to false.
-# For a public instance, this should probably be set to true.
-EMAIL_VERIFICATION_REQUIRED=FALSE
-
-# The URL of the application, used for generating links in emails
-APP_URL=http://localhost
-
-# The email address of the administrator of the system, who will receive notifications about new users, logins, etc.
-# Leave this empty if you don't want to receive any notifications
-ADMIN_EMAIL=
 
 ############################################################################################
 ### The values below should be left as is, unless you really need to use custom settings ###
@@ -80,14 +82,27 @@ ADMIN_EMAIL=
 APP_NAME=YAFFA
 
 # The environment in which the application is running
-APP_ENV=local
+# Set to "production" for general use, or "local" for development
+APP_ENV=production
+
+### Logging and debugging configuration ###
+# This is relevant, if you encounter any issues with the application, and you would like to see detailed error messages
+
+# Set this to true if you need to see detailed error messages
+APP_DEBUG=false
+
+# You can specify the log channel to be used for logging
+LOG_CHANNEL=stack
+
+# The minimum log level to be written to the log files
+LOG_LEVEL=error
 
 # Telescope should not be enabled, unless Composer dev dependencies are installed
 TELESCOPE_ENABLED=FALSE
 
 # Debugbar should not be enabled, unless Composer dev dependencies are installed
 DEBUGBAR_ENABLED=FALSE
-DEBUGBAR_OPEN_STORAGE=""
+DEBUGBAR_OPEN_STORAGE=
 
 ### Internal tools and services ###
 BROADCAST_DRIVER=log
@@ -99,7 +114,7 @@ REDIS_HOST=127.0.0.1
 REDIS_PASSWORD=null
 REDIS_PORT=6379
 
-# Sail settings
+# Laravel Sail settings
 WWWGROUP=1000
 WWWUSER=1000
 SAIL_XDEBUG_ENABLED=false
@@ -123,14 +138,14 @@ SENTRY_TRACES_SAMPLE_RATE=
 # Feel free to provide your own values, if you need any of these features
 
 # General flag to enable/disable some custom behavior or the Sandbox environment
-SANDBOX_MODE=false
+SANDBOX_MODE=FALSE
 
-# Recaptcha
+# Recaptcha - used to prevent spam and abuse on the login and registration forms
 RECAPTCHA_SITE_KEY=
 RECAPTCHA_SECRET_KEY=
 
-# Google Tag Manager
+# Google Tag Manager - used for tracking and analytics
 GTM_CONTAINER_ID=
 
-# CookieYes
+# CookieYes - used for cookie consent management, related to tracking and analytics
 COOKIEYES_ID=
