--- .env.example.docker
@@ -1,108 +1,0 @@
-# This file is a "template" of which env vars need to be defined for your application.
-# Copy this file as .env in the same directory with your docker-compose.yml file, and replace the values with your own
-# This version contains the minimal set of variables required to run your own YAFFA instance.
-# For a full list of available variables, see the .env.example.complete file
-
-### Basic app configuration ###
-
-# Encryption key used to encrypt session and other data
-# Make sure to keep this key secret, and don't share it with anyone.
-# Also, don't change it after you have started using the application.
-APP_KEY=PleaseReplaceThisWithARandomStringOf32Chars
-
-# Set this to true if you need to see detailed error messages
-APP_DEBUG=false
-
-# You can specify the log channel to be used for logging
-LOG_CHANNEL=stack
-
-# The minimum log level to be written to the log files
-LOG_LEVEL=error
-
-### Database configuration ###
-DB_CONNECTION=mysql
-DB_PORT=3306
-DB_DATABASE=yaffa_db
-DB_USERNAME=yaffa_user
-DB_PASSWORD=yaffa_password
-
-# Mail configuration
-# The current configuration uses Mailhog, which is a fake SMTP server that can be used for testing purposes
-# Your emails will be available at http://localhost:8025
-# You can also use a real SMTP servers, if you have one available, to actually send emails
-MAIL_MAILER=smtp
-MAIL_HOST=mailhog
-MAIL_PORT=1025
-MAIL_USERNAME=
-MAIL_PASSWORD=
-MAIL_ENCRYPTION=null
-MAIL_FROM_ADDRESS=info@yaffa.test
-MAIL_FROM_NAME=YAFFA
-
-### Configuration of external services ###
-# Your key for the Alpha Vantage API, which is used to fetch stock prices
-ALPHA_VANTAGE_KEY=
-
-MAILBOX_DRIVER=sendgrid
-INCOMING_RECEIPTS_EMAIL=
-MAILBOX_HTTP_PASSWORD=
-OPENAI_API_KEY=
-
-### Application-specific configuration ###
-
-# The maximum number of users that can be registered in the system.
-# For a personal instance, you can set this to 1 to prevent other people from registering.
-# Increase this value if you want to allow other users to register, or set it to empty to allow unlimited registrations.
-REGISTERED_USER_LIMIT=1
-
-# Does new users have to verify their email address before they can log in?
-# For a local, personal instance, without access to outgoing emails, this should be set to false.
-# For a public instance, this should probably be set to true.
-EMAIL_VERIFICATION_REQUIRED=FALSE
-
-# The email address of the administrator of the system, who will receive notifications about new users, logins, etc.
-# For Mailhog, this has not effect, but you can use your real email address if you use a real SMTP server
-# Leave this empty if you don't want to receive any notifications
-ADMIN_EMAIL=
-
-############################################################################################
-### The values below should be left as is, unless you really need to use custom settings ###
-############################################################################################
-
-# The name of the application, used in the UI
-APP_NAME=YAFFA
-
-# The environment in which the application is running
-APP_ENV=local
-
-# Name of the database container in docker-compose.yml
-DB_HOST=db
-
-# Telescope should not be enabled, unless Composer dev dependencies are installed
-TELESCOPE_ENABLED=FALSE
-
-# Debugbar should not be enabled, unless Composer dev dependencies are installed
-DEBUGBAR_ENABLED=FALSE
-
-# The URL of the application, used for generating links in emails
-APP_URL=http://localhost
-
-### Internal tools and services ###
-BROADCAST_DRIVER=log
-CACHE_DRIVER=file
-QUEUE_CONNECTION=database
-SESSION_DRIVER=file
-SESSION_LIFETIME=120
-REDIS_HOST=127.0.0.1
-REDIS_PASSWORD=null
-REDIS_PORT=6379
-
-# Recaptcha, only for Sandbox
-RECAPTCHA_SITE_KEY=
-RECAPTCHA_SECRET_KEY=
-
-# Google Tag Manager, only for Sandbox
-GTM_CONTAINER_ID=
-
-# CookieYes, only for Sandbox
-COOKIEYES_ID=
