+++ app/Providers/ViewServiceProvider.php
@@ -10,7 +10,8 @@ use App\Http\View\Composers\CurrencyListComposer;
 use App\Http\View\Composers\DataLayerEventComposer;
 use App\Http\View\Composers\InvestmentGroupListComposer;
 use App\Http\View\Composers\InvestmentPriceProviderListComposer;
-use App\Http\View\Composers\JavaScriptVariablesComposer;
+use App\Http\View\Composers\JavaScriptConfigVariablesComposer;
+use App\Http\View\Composers\JavaScriptUserVariablesComposer;
 use App\Http\View\Composers\NotificationMessageComposer;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
@@ -36,7 +37,11 @@ class ViewServiceProvider extends ServiceProvider
         View::composer('template.layouts.page', AccountListComposer::class);
 
         // Generic JavaScript variables
-        View::composer('template.layouts.page', JavaScriptVariablesComposer::class);
+        View::composer('template.layouts.page', JavaScriptConfigVariablesComposer::class);
+        View::composer('template.layouts.auth', JavaScriptConfigVariablesComposer::class);
+
+        // User-related JavaScript variables
+        View::composer('template.layouts.page', JavaScriptUserVariablesComposer::class);
 
         // Account form - all account groups
         View::composer('account.form', AccountGroupListComposer::class);
