+++ config/yaffa.php
@@ -11,7 +11,7 @@ return [
     'email_verification_required' => env('EMAIL_VERIFICATION_REQUIRED', false),
 
     // Optional settings, used primarily for the public facing Sandbox environment
-    'sandbox_mode' => env('SANDBOX_MODE', false),
+    'sandbox_mode' => boolval(env('SANDBOX_MODE',false)),
     'gtm_container_id' => env('GTM_CONTAINER_ID'),
     'cookieyes_id' => env('COOKIEYES_ID'),
 ];
