+++ resources/js/sandbox-components/login-helper.js
@@ -1,0 +1,8 @@
+const demoLoginButton =  document.getElementById('loginWithDemoCredentials');
+if (demoLoginButton) {
+    demoLoginButton.addEventListener('click', function () {
+        document.getElementById('email').value = 'demo@yaffa.cc';
+        document.getElementById('password').value = 'demo';
+        document.getElementById('login').click();
+    });
+}
