+++ resources/views/template/layouts/auth.blade.php
@@ -6,6 +6,7 @@
 @endsection
 
 {{-- Conditionally load Google Recaptcha --}}
+{{-- This feature is very closesly related to the sandbox mode, but it might be relevant to end-user instances, too --}}
 @if(config('recaptcha.api_site_key'))
 @section('head_scripts')
     <script src="https://www.google.com/recaptcha/api.js" async defer></script>
