+++ resources/views/template/sandbox-components/login-helper.blade.php
@@ -1,0 +1,15 @@
+<p class="text-medium-emphasis">
+    {{ __('sandbox.login.info') }}
+</p>
+<p class="text-medium-emphasis">
+    {!! __('sandbox.login.credentials') !!}
+    <button
+            class="btn btn-sm"
+            type="button"
+            title="{{ __('sandbox.login.fill') }}"
+            data-coreui-toggle="tooltip"
+            id="loginWithDemoCredentials"
+    >
+        <i class="fa-solid fa-right-to-bracket"></i>
+    </button>
+</p>
