+++ resources/views/template/sandbox-components/reset-alert.blade.php
@@ -1,0 +1,4 @@
+<div class="alert d-none d-md-block mb-0 ms-5 p-2 ps-3 pe-3 hidden" role="alert" id="sandBoxResetAlert">
+    <span></span>
+    {{ __('sandbox.reset.info') }}
+</div>
